+++ README.md
@@ -1,13 +1,14 @@
 # translate-mod
 
 将代码中的中文改成intl.get的形式
+使用jscodeshift作为转化插件
 
 ## 用法
 
 ```javascript
 script: {
   ...,
-  "translate": "translate build -p src",
+  "translate": "translate build",
   ...
 }
 ```
@@ -22,7 +23,11 @@ script: {
   // globby.sync options
   syncOptions: {
     ignore: `/.umi/**`, // /.umi/**
-    dot: true
-  }
+    dot: true,
+    ...
+  },
+  successCallBack: Function,
+  coreFn: (j: core.JSCodeshift, root: Collection<any>) => string,
+
 }
 ```
