+++ lib/transform.js
@@ -18,9 +18,14 @@ function mkDir(entry) {
   console.log(chalk.green('build successful'))
 }
 
-function transformer(content, parser) {
+function transformer(content, parser, coreFn) {
   const j = jscodeshift.withParser(parser);
   const root = j(content);
+  // 重构处理
+  return coreFn(j, root);
+}
+
+function coreFn (j, root) {
   const Literal = root.find(j.Literal, (p) => /[\u4e00-\u9fa5]/.test(p.value));
   const TemplateLiteral = root.find(j.TemplateLiteral);
   try {
@@ -90,9 +95,9 @@ function transformer(content, parser) {
   }
 
   return root.toSource();
+
 }
 
 module.exports = {
   transformer,
-  mkDir,
 };
